backend/requirements.txt
flask==2.0.3
flask-cors==3.0.10
python-dotenv==0.21.0
pytest==7.1.2
tox==3.27.1
